Simplify pipeline and test main
